chore(core): moving hasYes and removeCountryCode to a shared location (#8780)
* moving hasYes and removeCountryCode to shared

* adding spec file
