+++ libs/application/templates/announcement-of-death/src/lib/utils/index.ts
@@ -26,15 +26,3 @@ export const getFileRecipientName = (
   )
   return estateMember?.name || answers.applicantName.toString()
 }
-
-export const hasYes = (answer: any) => {
-  if (Array.isArray(answer)) {
-    return answer.includes(YES)
-  }
-
-  if (answer instanceof Object) {
-    return Object.values(answer).includes(YES)
-  }
-
-  return answer === YES
-}
