+++ libs/application/templates/driving-license/src/forms/draft/subSectionQualityPhoto.ts
@@ -5,6 +5,7 @@ import {
   buildRadioField,
   buildSubSection,
   getValueViaPath,
+  hasYes,
 } from '@island.is/application/core'
 import { m } from '../../lib/messages'
 import { HasQualityPhotoData } from '../../lib/types'
@@ -12,7 +13,6 @@ import { NO, YES } from '../../lib/constants'
 import { B_FULL } from '../../shared/constants'
 import {
   hasNoDrivingLicenseInOtherCountry,
-  hasYes,
   isApplicationForCondition,
   isVisible,
 } from '../../lib/utils'
