+++ libs/application/templates/driving-license/src/lib/utils/index.ts
@@ -1,2 +1,1 @@
-export * from './hasYes'
 export * from './formUtils'
