--- libs/application/templates/estate/src/lib/utils/index.ts
@@ -1,13 +1,0 @@
-import { YES } from '../constants'
-
-export const hasYes = (answer: any) => {
-  if (Array.isArray(answer)) {
-    return answer.includes(YES)
-  }
-
-  if (answer instanceof Object) {
-    return Object.values(answer).includes(YES)
-  }
-
-  return answer === YES
-}
