--- libs/application/templates/passport/src/utils/removeCountryCode.ts
@@ -1,3 +1,0 @@
-export const removeCountryCode = (phone: string) => {
-  return phone.replace(/(^00354|^\+354|\D)/g, '')
-}
